Bill Book Changes

+++ src/main/java/com/juri/printing/dao/Prnt08PrintDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.juri.printing.entity.Prnt08Print;
 
 @Repository
-public interface Prnt08PrintDao extends JpaRepository<Prnt08Print, Integer> {
+public interface Prnt08PrintDao extends JpaRepository<Prnt08Print, Long> {
 
 	@Query("select e from Prnt08Print e where e.prnt08PrintType = :printName and e.prnt08Range = :range")
 	Prnt08Print findByPrnt08PrintTypeAndPrnt08Range(@Param("printName") String printName,@Param("range") String range);
